ENH: Enable internationalization
Build the application with internationalization enabled (Slicer_BUILD_I18N_SUPPORT=ON) by default and if internationalization is not explicitly disabled in application settings (using Internationalization/Enabled) then the language selector box is displayed in application settings.

The commit also slightly improves discovery of translation (*.qm) files: the search path is <SlicerHome>/bin/translations (both build tree and install tree); and in addition to that in the install tree also folders listed in Slicer_QM_OUTPUT_DIRS are searched for translations.

Updated CTK to improve the language selector combobox.