ENH: Flip 3D view mousewheel zoom direction for consistency with slice views
It would be nice to have consistent zoom direction of Ctrl-wheel in slice view and wheel in 3D view. The zoom direction of wheeling was chosen to be consistent with the right-click-and-drag action: "pulling back the finger on the wheel" and "pulling back the whole mouse while holding right button down" both zooms in the view. This slight inconsistency is present in slice views as well and people don't complain (although Ctrl-wheel is probably much less used than simple wheel), so it should be acceptable for 3D views as well.

We could consider flipping the right-click-and-drag zoom direction to be consistent with the mouse wheel direction. However, the mouse range of motion from neutral position to forward is about 3x shorter than from neutral position to backward, so inverting the zoom would make it much harder to quickly zoom in to see some small detail and the zoom back. It is also a very commonly used gesture and users would need to re-train muscle memory.

So, flipping mousewheel zoom direction in 3D views and adding Ctrl-mousewheel action (for pinch-to-zoom on Windows touchpads) seem to be the right things to do.

Co-authored-by: Pierre BOUR <riep@users.noreply.github.com>