ENH: Add setMRMLScene to qMRMLCollapsibleButton
This adds the setMRMLScene function to `qMRMLCollapsibleButton`, as well
as the qt signaling to propagate the changes of the MRMLScene.

This is useful to propagate the change of MRMLScene to children of `qMRMLCollapsibleButton`,
which is particularly useful when defining signal/slot arrows in Qt designer. It replaces many long
signal/slot arrows from the top-level widget to each child MRML widget in the collapsibleButton
with short signal/slot arrows from the collapsibleButton to each child MRML widget inside it.