BUG: Export DICOM image into the selected study
Study instance UID was saved with an incorrect name (without the "DICOM." prefix) into the study item in the subject hierarchy.
This made the tag in the exportable be empty by default and so a new study was created each time.

This issue was fixed by fixing saving of the Study instance UID value.
Since GDCM can only accept all or nothing for UIDs, implemented automatic generation of series instance UID and frame of reference UID (used when these UIDs are not specified).

Additional fixes:
- Fixed a typo in a tag name: "Frame of reference UID" was written incorrectly as "Frame of reference instance UID". The old tag name is still supported in DICOMExportScalarVolume script for backward compatibility but it is deprecated now.
- Improved error reporting in DICOM examine and export: if an exception is thrown then an error message is displayed and details of the exception are added to the application log.