ENH: Add startupWorkingPath property to qSlicerCoreApplication
The current working directory can be changed by any module at any time (for example, because some algorithms or executables use the current working directory for their inputs or outputs).
Therefore, modules that wanted to use the current working directory to interpret command-line arguments, or the Slicer Jupyter kernel that relies on knowing the working directory could not reliably get it.

A new property `startupWorkingPath` was added to qSlicerCoreApplication. This property is initialized from the current working directory at startup and cannot be changed later, therefore it is a reliable way of getting the original working directory.

The name is derived from "current working directory", with "current" replaced by "startup" (similarly to `startupEnvironment` property) and "directory" replaced by "path" (as all other directory path strings are referred to as `...path` in qSlicerCoreApplication and in Qt in general).