COMP: Remove unused vtkPVScalarBarActor
This actor is not used anymore in Slicer core and any of the public Slicer extensions, therefore it is removed.
Since this was the only reason why vtkSlicerDataProbeModuleLogic library existed, the library was removed as well.
This reduces the amount of code and will make it easier to refactor remaining features of the DataProbe in the future.