BUG: Fix Qt Designer crash
`qSlicerApplication::application()` may return nullptr (it happens when the Qt application is not a qSlicerApplication) but the return value was not checked.
When Qt Designer application instantiated the qSlicerTerminologyNavigatorWidget then nullptr was returned and used, causing a crash.

Fixed by adding a null-pointer check.

fixes https://github.com/Slicer/Slicer/issues/6167