ENH: Add Markups Point List Node
This is a renaming of a Markups Fiducial node to a Markups Point List node. A Fiducial node is still maintained for backwards compatibility, but can be removed at some point in the future. A vtkMRMLMarkupsFiducialNode is now a derived class of vtkMRMLMarkupsPointListNode to make it easier to cut off and remove the fiducial node class.

This is a continuation of https://github.com/Slicer/Slicer/commit/8e294598acddf71414bc2f1c2ff7c67b2e871683 which updated the create a fiducial node button to be a create a point list node button.