ENH: Add options to toggle individual markups interaction handle axes
Previously, all of the interaction handles could only be enabled/disabled with others of the same type (rotation, translation, scaling).
This commit adds options for enabling disabling the individual XYZ axes for each of the handle types, as well has the handles that allow free transformation.

ROI nodes now display translation handles by default, but only display the free translation point in the center of the ROI.

Closes #5558
Re #5061