BUG: Fix saving duplicate states in vtkSegmentationHistory
When save state is requested in vtkSegmentationHistory then the history must contain the current state and all previous states.
This means that if save state is requested when the current state is already stored in the history then there is no need to save the state again
(it would just result in having two identical states in the history).

When a state is saved we also set the current state as last restored state. This avoids creation of an extra (identical) saved state when SaveState is called but the segmentation is not modified.

Fixes #6183