ENH: Add saved scene filenames to File menu's Recent list
Saving the last saved scene file (.mrml or .mrb) makes it easy to reopen saved scenes after restarting Slicer.

Other saved files are not added to the recent file list, because a scene can contain dozens of nodes and adding all of them to the recent list would immediately discard all previous items in the list.
If a module wants to add a file to the list then qSlicerMainWindow::addFileToRecentFiles method can be used.