COMP: Remove ExpertAutomatedRegistration module
ExpertAutomatedRegistration module was added as a simple example for showing how ITK can be used in Slicer CLI modules.
Registration quality and configurability of this module is very limited compared to BRAINS, Elastix, and ANTs, which are all available in Slicer.
By removing it avoids user confusion (as the module is not recommended to be used) and reduces maintenance burden for developer.

The module was marked for removal several years ago, but it is a good time to retire it now, as it is incompatible with latest ITK version that Slicer has upgraded to.