ENH: Make pip_install more robust
`slicer.util.pip_install` uses `slicer.util.logProcessOutput` to run the process and log the output.
In case the system code page is not UTF-8, the process output may not be converted to UTF-8 if it contains non-ASCII character, making the process execution return with failure.

Since missing non-ASCII process output (typically the progress bar) is generally tolerable, this commit updates `slicer.util.logProcessOutput` to tolerate UTF-8 conversion errors (can return with success even if some parts of the process output was not convertible to UTF-8).