BUG: Fixed application window not clickable when popup is displayed
On Windows, some popup windows (Add data, Settings, Extensions manager, etc.) prevented switching to the application. When the user clicked on the application window, the application window or the popup was not raised to the top.

How to reproduce:
- Open "Add data" window
- Switch to another application that covers part of the Slicer application window
- Click on the application window (outside the "Add data" window) => ERROR: the application window is not brought to the top.

The problem was that the parent of "Add data" window (and a number of other popups) did not set the application main window as parent. On Windows, modal popups prevent processing events in all other windows, therefore clicking on the application main window has no effect, unless it is parent of the popup window.

Fixed the issue by setting the main window as parent for all modal popups.