ENH: Make additional options in "Add data" more visible
When user checked "Options" checkbox in "Add data" then the additional options was added to the window but the user had to scroll to the right to see them.

This commit scrolls the window automatically so that the options are visible.
In addition, the column sizes are optimized so that the filename column does not completely pushes out the options column from the view.