BUG: Fix crash in vtkMRMLSliceLayerLogic::UpdateNodeReferences
When volumes were loaded in quick succession (e.g., load multiple images using Sample Data module) then in some cases vtkMRMLSliceLayerLogic::UpdateNodeReferences crashed due to this->VolumeNode being nullptr in the line `this->VolumeNode->SetAndObserveDisplayNodeID(displayNode->GetID());`.

The only way it could happen is that the some module or GUI element set this->VolumeNode member variable to nullptr, so while the pointer was check to be valid at the beginning, it was not valid anymore when being used later.

The issue was fixed by using the pointer that was obtained at the beginning of the method.