COMP: Fix install of numpy on macOS 10.13 updating from 1.22.2 back to 1.22.1
This commit partially reverts d663399b0 (COMP: Update python packages to
latest) to fix installation of numpy on macOS 10.13 (corresponding to
the version of macOS used on the "factory-south-macos" builder).

For consistency, the version of numpy is reverted back to 1.22.1 for
all platforms.

This change is required because starting with numpy 1.22.2, the minimum
version of macOS supported is 10.14.

See MacPython/numpy-wheels@b61edce01 (Update Mac version from 10.9 to 10.14.)
and https://github.com/MacPython/numpy-wheels/runs/5008197518 for more
details about the error requiring the update to 10.14 in version 1.22.2.

Since scipy 1.8.0 works with numpy >= 1.17.3 and < 1.25.0, reverting the
version should not introduce incompatibilities.
See https://github.com/scipy/scipy/blob/v1.8.0/setup.py#L544-L553

Co-authored-by: James Butler <jbutler@sonovol.com>