BUG: Fix interpolation reset issue in slice views
When the interpolation was different for the previous and new (background or foreground) volume in the slice views when changing, the previous volume display node was set the interpolation value of the new volume display node. It was because of a signal that was emitted when setting the interpolation on the checkbox in the slice view controller, which triggered setting that interpolation value. By adding blockSignals calls this issue is fixed.