BUG: Fix markups glyph scaling
Markups glyph size was smaller on HiDPI screens due to a VTK bug and so it was not possible to
choose a default value that works well for all kinds of screens.

The VTK bug was fixed in VTK (https://gitlab.kitware.com/vtk/vtk/-/commit/26fc4d9dc9780caa66b8de2f8a474e21f5975fc0)
and therefore the glyph scale is now set to 3%, which works reasonably well on all screens.

fixes #5180