ENH: Move "Application settings" action to the top of Edit menu
Edit menu's Cut/Copy/Paste actions are not useful (with a few exceptions, copy-paste is only supported in textboxes where keyboard shortcuts and right-click menu are more convenient to use).
It is fine to have them there to make menu sizes a bit more balanced.

In contrast, "Application Settings" action is used often and it is much harder to click at the bottom of the menu than at the top.

Therefore, this commit moves "Application Settings" action to the top of the Edit menu.