COMP: Remove unused Spline Annotations module MRML node
This commit removes support for Spline node (class vtkMRMLAnnotationSplineNode).
The corresponding functionality has never been available to the user since
it was originally introduced in 2011.

The support for placing a new Spline has never been enabled in the
vtkSlicerAnnotationModuleLogic class.

See 38cb7b681 (ENH: add to the list of annotation ids that can be active,
with icons. Listen for the control point modified event)