BUG: segfault invoked by SettingsDialog deletion
Previous impl-n worked only when when mainwindow is deleted after app.
If mainw-w is deleted before app then we get seg fault in destructor.
I.E. we are trying to set parent to destroyed SettingsDialog.
This used to happen with SlicerCAT.

QPointer helps to track the lifetime of a pointer.
So we always know if it is alive.

Steps to reproduce:
1) launch SlicerCAT from command line
1) open `Application settings`
2) close it
3) close application
4) see output to command line