BUG: Segfault at application exit due to SettingsDialog deletion
Previous implementation worked only when when mainwindow is deleted after app. If mainwindow was deleted before app then we got seg fault in application destructor (due to trying to set parent to destroyed SettingsDialog). This used to happen with SlicerCAT.

Steps to reproduce:
1) launch SlicerCAT from command line
1) open `Application settings`
2) close it
3) close application
4) see output to command line

Fix: Use QPointer to track the lifetime of the settings dialog. If the main window has already deleted the settings dialog then the application does not delete it. If the settings dialog has not been deleted yet then the application deletes it and also sets the parent to nullptr to ensure the main window will not attempt to delete it.