ENH: Update BRAINSTools GitHub links with new main branch name
BRAINSTools has renamed the default branch name from master to main.

This is for clarity because this is technically not necessary. "When you rename a branch on GitHub.com, any URLs that contain the old branch name are automatically redirected to the equivalent URL for the renamed branch."

Source: https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/managing-branches-in-your-repository/renaming-a-branch