BUG: Set markups position status to defined by default
If markups file is generated manually then control point position may be defined without positionStatus set to defined.
This was interpreted as undefined position, which was unexpected/inconvenient (see https://discourse.slicer.org/t/using-json-files-to-create-markups/22754).

This commit changes the behavior so that if control point "position" is specified and "positionStatus" is not specified in the file then "positionStatus" will be considered as "defined".