BUG: Fix crash of Level Tracing effect on rotated slice views
Level tracing filter only works for non-oblique slices.
When attempting to use the effect on oblique slices (that are not aligned to the slice view axis) then the seed point could get out of range, which raised an ITK exception, which resulted in providing empty inptu to the TransformPoints function in VTK, which then crashed.

Fixed the issue by displaying an error if the slice view is rotated and not try to trace.
Also added checks in the segment editor effect to ensure that failed tracing does not crash the application.