ENH: Improve python_package_updater to validate hints
Pair of hints are expected to be specified as CMake comment using the following
convention where `<auto-generated-requirements>` is replaced by the
`python_package_updater` script.

  # [packagename]
  <auto-generated-requirements>
  # [/packagename]

This commit updates the script to always perform validation and exit
if there are errors.

The option "--validate" was also added to only perform validation without
updating external projects.

List of errors are formatted like this:

  Searching external projects in /home/jcfr/Projects/ShapeVariationAnalyzer/SuperBuild
  Validate external projects
  /home/jcfr/Projects/ShapeVariationAnalyzer/SuperBuild/External_python-ShapeVariationAnalyzer-requirements.cmake
    joblib != packaging

Co-authored-by: James Butler <jbutler@sonovol.com>