ENH: Add "Show 3D" button to segmentation module
When a segmentation was loaded and the user did not want to edit the segmentation, only show it in 3D it was not intuitive that the user had to go to Segment Editor module (or click on Create button of Closed surface representation).

Implemented by creating a qMRMLSegmentationShow3DButton widget for showing/hiding the segmentation in 3D.
It is used now in both Segmentations and Segment Editor module and can be used in custom modules, too.