BUG: Fix hardening transform on volume nodes
When an all-linear vtkGeneralTransform was applied on a volume node then the transform was considered as non-linear.
Fixed this by inspecting content of vtkGeneralTransform when checking if a transform is linear.

When non-linear transform was applied to a volume then the extent was not computed correctly.
Fixed by setting the origin and extent correctly.