BUG: Fix slice views remaining black after loading a scene
Initialization of views was incomplete when the corresponding view nodes were already in the scene, but the view widgets were not instantiated yet; and a scene was loaded that had the non-instantiated view widgets marked as "MappedInLayout".

Fixed by not restoring the "MappedInLayout" state of views from the saved scene. It makes sense not to restore MappedInLayout state from saved scene, because the fact that a view was mapped in layout when the scene was saved does not mean that the view is mapped in layout at the time of loading the scene.

Fixes #6284