BUG: Allow slicer.util.pip_install to take input arguments with spaces
When a package was attempted to be installed from a folder that contained spaces then pip_install failed becaue the folder name was split to multiple arguments:

    PythonSlicer.EXE', '-m', 'pip', 'install', 'C:/Users/Benjamin', 'Colbert/AppData/Local/NA-MIC/Slicer/cache/SlicerIO/open3d-0.14.1+816263b-cp39-cp39-win_amd64.whl']' returned non-zero exit status 1.

Fixed the issue by using shlex.split (that does not split quoted strings) instead of simple string split. Also added an option to specify pip_install and pip_uninstall arguments as a list, to allow users to avoid the need to quote strings.

Fixes the issue described at https://discourse.slicer.org/t/installing-open3d-is-giving-an-error/17978/9