BUG: Fix saving of scalar volume sequence with only one frame
When trying to save a volume sequence with a single frame, vtkTeemNRRDWriter raises the following error:
```
[nrrd] nrrdSave:
[nrrd] nrrdWrite: trouble
[nrrd] _nrrdWrite:
[nrrd] nrrdCheck: trouble
[nrrd] _nrrdCheck: trouble with space field
[nrrd] _nrrdFieldCheck_space: trouble
[nrrd] _nrrdFieldCheckSpaceInfo: axis[0] has a direction vector, and so can't have min, max, spacing, or units set
```

This is because the range dimension is omitted by the writer when the size of the first dimension is 1. This results in an error when units are instead assigned to the first spatial dimension.
Fixed by adding an option (vtkTeemNRRDWriter::ForceRangeAxis) to force the writer to include a range axis, even when the size of the first dimension is 1. This option remains disabled by default, and is only enabled by vtkMRMLVolumeSequenceStorageNode.

This commit also adds code to vtkMRMLSequenceStorageNodeTest1.cxx to test reading/writing sequences.