BUG: Fix remote cache directory in vtkMRMLSequenceStorageNodeTest1
Because the remote cache directory was not specified for the scene in vtkMRMLSequenceStorageNodeTest1, attempting to unzip and read the seq.mrb would fail on Linux and Mac as the application would not have permission to write to the directory.

Fixed by setting the remote cache directory to the temporary testing directory.