BUG: Fix copying markup measurements
When markup nodes were copied then custom measurements (that are not added in the markup node constructor) were not copied.

Fixed by adding a method that can create a new instance of a measurement and use that to create a copy of every measurements when a markup node is copied.

Note that VTK_NEWINSTANCE is specified for the new vtkMRMLMeasurement::CreateInstance method. This has an effect when using CreateInstance in Python: it is not necessary to call `UnRegister` on the object that `CreateInstance` returns.

Changed vtkMRMLMeasurement::ControlPointValues to be a vtkSmartPointer for safer management of the pointer.