BUG: Fix infinite loop when loading singleton transforms
(cherry picked from commit 5514b17f8d)

When loading a scene that contains singleton transform nodes it is possible to enter an infinite loop calling vtkMRMLTransformNode::GetTransformBetweenNodes(). This can happen if a circular loop is created between the transform nodes as their references are being updated.

Fixed by checking for duplicate transform nodes in vtkMRMLTransformNode::GetTransformBetweenNodes if the maximum transform depth is reached (100). If a loop is detected, then the transform is set to identity, and a warning is logged.

Re #6355