ENH: Increase maximum number of colors loaded from colortable files
File loading performance is fixed by implementing bulk setting of color values.

Increased maximum number of accepted entries to 1 million as requested in #6354.
Loading of a file with 1 million as the maximum ID takes only a few seconds.
Colors module performance is not great with such huge tables. It takes about 20 seconds
to display a color table with 1 million entries.

fixes #6354