COMP: Do not require building extension from git repository
(cherry picked from commit 1a0daf931)

Developers often complained that their extension build failed for some
unknown reason and frequently the issue was that they did not store the
source code in a git repository.

It is good to encourage developers to use best practices, such as version
control, but making the build fail if they build from a non-version-controlled
repository was unnecessarily strong measure.

This commit changes the behavior so that if an extension is built from a
non-git folder then the build succeeds, just a warning is displayed during
project configuration step and `NA` is used as the extension's WC revision.

Co-authored-by: Andras Lasso <lasso@queensu.ca>