BUG: Fix loading of table node without schema
(cherry picked from commit 260478f415)

Repeated loading of empty table node failed. The issue was that for empty tables no schema file was generated and the schema filename was set to "".
However, when the scene was loaded this empty filename was converted to an absolute path (a folder name). When this scene was saved and loaded again then loading the scene file from that path failed.

Fixed the issue by leaving empty filenames empty during scene loading (not converting them to an absolute path).

Fixes #6371

Co-authored-by: Gabriel Muller <gabriel.muller@ltrace.com.br>