PERF: Improve markups control point movement speed
(cherry picked from commit 423a07be7)

While a markup control point is moved then the displayable managers keep checking if the active place node class is supported.
It was checked by creating a node object each time, which was an expensive operation.

Improve performance by determining if the active place node class is supported by checking if it is registered in the markups logic.