BUG: Fix jumping to locked markup control point
(cherry picked from commit c79bded940)

When a control point is locked it means that its position cannot be changed by click-and-dragging in a view.
However, we still want to be able to click on a control point to jump to that position in all views.

Complete lock of all interactions with the markup can be achieved by locking the node instead of locking specific control points.