BUG: Make qSlicerUtilsTest1 more robust
(cherry picked from commit 986744629)

When the working copy is reconfigured and a full Slicer build is not performed then qSlicerUtilsTest1 failed with:

Line 408 - Problem with isPluginBuiltIn(). Expected 0 for:
	path: /Applications/Slicer.app/Contents/Extensions-30961/plugin.txt
	applicationHomeDir: /Applications/Slicer.app

It was because Slicer revision in vtkSlicerConfigure.h and vtkSlicerVersionConfigure.h were different.

Made the test more robust by taking the version information from Slicer_REVISION, which is the same as the tested
vtkSlicerApplicationLogic::IsPluginBuiltIn method uses.