ENH: Optimize processing Segment Editor updates during Batch Processing
For example when removing volume nodes, this can trigger a change of the current Source Volume node specification in the Segment Editor module if the current volume node is removed from the scene. When this happens over and over during batch processing there was frequent updating of the source volume node for SegmentEditor logic which can be an expensive operation, especially the sourceVolumeNodeChanged call for SegmentEditorThresholdEffect.