BUG: Fix crash in interactive slice intersections
(cherry picked from commit 197a96b2c)

When a view contained markups and interactive slice intersections, the application could crash while moving and dragging markup control points.

The problem was that when a widget was forced to release the focus the `widget->Leave(eventData)` method was called with `nullptr`, but the `Leave`
method actually required valid `eventData`.

Fixed the issue by passing valid `eventData` to `SetHasFocus` method, which then can call `Leave` with that `eventData`.

fixes #6400