ENH: Improve performance of getting extensions install history
(cherry picked from commit 8c148a87f)

The extensions server was queried for metadata of all previously installed extensions for all versions.
This time could add up to several seconds (or even tens of seconds on a computer with very long history of Slicer installations).

This commit changes the behavior that the extensions server is only queried for the latest version for each previously installed extension.

Also fixed ordering of revision numbers. Now numeric sorting is used instead of lexicographic.