BUG: Fix VTKObservationMixin preventing observed nodes from being released
Fixes #6406.

This change modifies the implementation of `VTKObservationMixin` to use a
'weakref.WeakKeyDictionary` rather than a flat list of registered
observations. This allows observed vtkObjects to be garbage collected if
all other references have been lost.

The registered events are stored in a nested dictionary structure to
acheive constant-time access for `removeObserver`, `hasObserver`, etc.
This also guarantees no observations are duplicated.

```
{obj: {event: {method: (group, tag, priority)}}}
```