DOC: Update build instructions to include CentOS 7
Slicer built on CentOS 7 will be available for many Linux distros.

At the time of this commit, Slicer was successfully build, packaged
and ran on:
* Debian 10.12
* Debian 11.3
* Ubuntu 18.04
* Ubuntu 20.04
* Ubuntu 22.04
* Fedora 34

To support running on Debian 10.12, the "Getting Started" section
is also updated to document tricks needed to solve `xinerama` issue.

See https://discourse.slicer.org/t/build-and-package-slicercat-on-manylinux/23443/3

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>