ENH: Remove custom itkDCMTKFileReader from PETStandardUptakeValueComputation
(cherry picked from commit 68373c19b)

itkDCMTKFileReader was added in 2012 but it seems that current itkDCMTKFileReader in ITK works just as well.
The 2012 version has a bug (it should use (0x5200, 0x9230) instead of (0x5200, 0x9239) tag - see https://discourse.slicer.org/t/image-position-patient-attribute-0020-0032-in-dicom-file-not-handled/23950/3) therefore it makes sense to switch to the ITK implementation.