BUG: Fix compatibility of Markups JSON with Slicer < 5.1
Prior to recent changes, Slicer expected a specific schema url, which is different from the one that is currently used.

- Slicer <= 5.0.2:
https://raw.githubusercontent.com/slicer/slicer/master/Modules/Loadable/Markups/Resources/Schema/markups-schema-v1.0.3.json#"

- Current version of Slicer:
https://raw.githubusercontent.com/slicer/slicer/main/modules/loadable/markups/resources/schema/markups-schema-v1.0.3.json#

Fixed compatibility issues by using the old url. In the future, we will update the schema url to reflect the current version of Slicer.

This commit also relaxes the regex url validation to only check the filename and ignore the rest of the url.