DOC: Add Contributor Covenant Code of Conduct v2.1
While reviewing project metrics for a grant application we realized that
although we have a common agreement about how to conduct operations in the
Slicer community, we do not have a formalized document to describe it.

The Contributor Covenant is a short and clear description that captures
well how we operate, so we propose to adopt this as code of conduct document.

The text was copied verbatim from [1] and is indented to be updated in
follow-up commits. This will allow to clearly track which changes were
integrated by reviewing the file history at [2].

[1] https://www.contributor-covenant.org/version/2/1/code_of_conduct/code_of_conduct.md
[2] https://github.com/Slicer/Slicer/commits/main/CODE_OF_CONDUCT.md

Co-authored-by: Andras Lasso <lasso@queensu.ca>
Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillion@kitware.com>
Co-authored-by: Sonia Pujol <spujol@bwh.harvard.edu>
Co-authored-by: Steve Pieper <pieper@isomics.com>