ENH: Display git diff when extension update is available
If an extension update is available and the extension is hosted on github or gitlab (currently all extensions are hosted on one of these services)
then a link is generated that allows users and developers to see what exactly has changed.

Also updated the extension manager to set the focus to the search box by default so that users can find extensions just by start typing
(without the need to click in the searchbox).