BUG: Restore widget name in qMRMLSegmentEditorWidget
This commit fixes a regression introduced in
1b94b0d53 (ENH: Improve Segment Editor widget layout)

It restores the object name from `Form` to `SegmentsTableResizableFrame`.

Using an explicit widget name is relevant in the context of prototyping
and/or custom application development.

Additionally, it adds the missing `customwidget` definition
associated with the `ctkMenuButton` custom widget.
Sew https://doc.qt.io/qt-5/designer-creating-custom-widgets.html

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>