ENH: Reorder favorite modules by drag-and-drop
Favorite modules (in Application settings / Modules) can now be reordered by drag-and-drop.
When a module is drag-and-dropped into the favorite list, the drop position is used (instead of adding the module to the end of the list).
Adding/removing/reordering a module shortcut takes effect immediately (no need to restart the application).

fixes #6511
fixes #6510