ENH: support markups with distorted zoom
Fix markups repr-n when camera non-equally scaled along different axes.
Markups is designed to be always oriented away from camera.
Orientation is done with `GlyphOrientationArray`.
Scaling is designed to be similar to orientation.
With Glyph Mapper `SetScaleModeToScaleByVectorComponents`.
As orientation is updated with `UpdateControlPointGlyphOrientation`,
Scaling is updated with `UpdateControlPointGlyphScale`.
Now it is a draft as I can't correctly project scaling from
global coords to markups point coordinates.