BUG: Make node names and quantity identifiers non-translatable
In Qt .ui files, strings and stringlists are marked as translatable by default, which makes MRML node names and quantity identifier show up in translation source (.ts) files as a translatable string. This may be confusing for translators and if the strings are actually translated then it can cause issues.

Fixed the issue by marking all existing node class string lists and quantities as non-translatable.
As a preventive measure, updated all custom widget plugins so that non-translatable string and stringlist properties are set to non-translatable by default.

Note: Qt has the option to prevent string translation options to show up in Qt designer by making domXml() method return stringpropertyspecification like this:

  return  "<ui language=\"c++\">\n"
    // Set default widget property attributes
    "<widget class=\"qMRMLCoordinatesWidget\" name=\"MRMLCoordinatesWidget\">\n"
    "</widget>\n"
    "<customwidgets>\n"
    "  <customwidget> <class>qMRMLNodeComboBox</class>\n"
    "    <propertyspecifications>\n"
    "      <stringpropertyspecification name=\"quantity\" notr=\"true\" type=\"singleline\" />\n"
    "    </propertyspecifications>\n"
    "  </customwidget>\n"
    "</customwidgets>\n"
    "</ui>\n";

However, we did not choose to use this mechanism because it has two limitations:
- it does not prevent non-translatable strings from showing up in .ts files
- it only works for string properties (not for stringlist)