ENH: Fix use of run_ctest_ variables
This commit includes a new check based on the `run_ctest_submit`
variable when submitting `PARTS build`. Without this fix, you could try
to disable submission to dashboard but the build information will still
appear submitted.

This commit modifies the logic of the script to decouple the building of
the Slicer package and its upload. This is to support cases where the
Slicer package needs to be built, but won't be uploaded. Not having this
in place forces the generated package to be uploaded.

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>