BUG: Minor VectorToScalarVolume module improvements
When scene was closed or the module was reloaded then the module threw exceptions.
Fixed by updating the module to use the latest scripted module template.

Average computation was not correct for integer volumes. For example for a volume containing
30, 50, 100 in 3 channels the average is 60 but it was computed to be 59 (due to casting to int).
Fixed by rounding of the average computation result instead of just casting the floating-point value.

fixes #6478