BUG: Fix saving of sequence nodes
Saving of sequence node was successful the first time but failed on subsequent saves.

Root cause of the problem: A vtkMRMLSequenceNode singleton node was added to the internal "sequence scene" to store the index values. When the node was saved again, the singleton node already was in the scene and so Copy method was used to update it. However, Copy method did not work correctly because this internal sequence node was incomplete (only stored index values, not node IDs).

Fixed the issue by:
- removing the temporary internal vtkMRMLSequenceNode so that copy of the node is not attempted
- making vtkMRMLSequenceNode copying work if the node does not store data nodes (just node IDs) to make it possible to copy incomplete sequence nodes
- add reporting of errors or warnings occur in data node's WriteXML method calls to prevent silent failure of node saving