ENH: Make it easier to export multiple series under the same DICOM study
When attempting to export a data node that is not in a DICOM hierarchy (patient/study) then Slicer offered to create these items automatically.

Previously, the items were not initialized with any IDs and instead the IDs were generated in the exported files.
However, this made all the exported items have a different patient and study instance.
If the user wanted to add more series to the study then the user either had to manually specify IDs or had to load the data sets from the DICOM database to get the automatically generated IDs.

This commit changes the behavior so that the automatically created patient and study items have valid IDs in them, making it easy to export multiple series under the same DICOM patient/study.