ENH: Simplify scene export to Slicer Data Bundle DICOM file
Previously a reference DICOM file was needed for generation of a Slicer Data Bundle (SDB) DICOM file.
If none of the data in the scene was loaded from DICOM then no such reference file existed, which required the user to create one (for example, by exporting an image to the DICOM database). The need for this extra step was complicated and non-intuitive for users.

This commit allows exporting an SDB DICOM file without a reference DICOM file. A simple template from DCMTK is used as a basis for the export then.