ENH: Avoid warnings when adding nodes to the scene
Whenever a new node was added to the scene, all the "pre" and "post" extra items of the scene model were removed and re-added in all the scene models (in all node selector comboboxes, tree views, etc.). In some cases it led to temporarily selecting the just added MRML node, which generated warnings if the node had not been completely initialized with data yet.

fixes #6098