BUG: Fix crash when exporting segment model node
Exporting visible segments to model nodes could cause a crash when exporting using the action on the virtual "Segment" items, but wouldn't crash when exporting from the "Segmentation" item. When exporting from segment, the segment parent (the segmentation node) was used as the parent of the folder node. This seems to cause an invalid SH item parent, and could cause a crash later when the model is updated.

Fixed by using the parent of the Segmentation node for the export folder parent.

Fixes #6570